# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/admin/users": {
            "post": {
                "description": "create user and admin , admin only",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "create users",
                "parameters": [
                    {
                        "type": "string",
                        "description": "unique email",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "password that have at least 8 length and contain an alphabet and number ",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "repeatPassword that have at least 8 length and contain an alphabet and number ",
                        "name": "repeatPassword",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "firstName",
                        "name": "firstName",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "lastName",
                        "name": "lastName",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "isAdmin",
                        "name": "isAdmin",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.UsersListResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/swagger.UnauthenticatedResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/swagger.AccessForbiddenResponse"
                        }
                    }
                }
            }
        },
        "/admin/users/{id}": {
            "delete": {
                "description": "delete user or admin , admin only",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "delete user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "user id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.UsersListResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/swagger.UnauthenticatedResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/swagger.AccessForbiddenResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/swagger.NotFoundResponse"
                        }
                    }
                }
            }
        },
        "/auth/access-token-verify": {
            "post": {
                "description": "jwt access token verify",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "access token verify",
                "parameters": [
                    {
                        "type": "string",
                        "description": "accessToken",
                        "name": "accessToken",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.SuccessResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/swagger.FailedResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/swagger.FailedValidationResponse"
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "description": "jwt login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "login",
                "parameters": [
                    {
                        "type": "string",
                        "description": "email",
                        "name": "email",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "password",
                        "name": "password",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.LoginResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/swagger.FailedLoginResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/swagger.FailedValidationResponse"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "jwt register",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "register",
                "parameters": [
                    {
                        "type": "string",
                        "description": "unique email",
                        "name": "email",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "password that have at least 8 length and contain an alphabet and number ",
                        "name": "password",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "repeatPassword that have at least 8 length and contain an alphabet and number ",
                        "name": "repeatPassword",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "firstName",
                        "name": "firstName",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "lastName",
                        "name": "lastName",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.SuccessResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/swagger.FailedValidationResponse"
                        }
                    }
                }
            }
        },
        "/auth/renew-access-token": {
            "post": {
                "description": "jwt renew access token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "renew access token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "refreshToken",
                        "name": "refreshToken",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.SuccessVerifyAccessTokenResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/swagger.FailedResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/swagger.FailedValidationResponse"
                        }
                    }
                }
            }
        },
        "/ping": {
            "get": {
                "description": "do ping",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "generic"
                ],
                "summary": "ping",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.PingResponse"
                        }
                    }
                }
            }
        },
        "/users": {
            "get": {
                "description": "list of paginated response , authentication required",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "get users list",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.UsersListResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/swagger.UnauthenticatedResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/swagger.AccessForbiddenResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "DTO.AccessTokenRes": {
            "type": "object",
            "required": [
                "accessToken",
                "expAccessToken"
            ],
            "properties": {
                "accessToken": {
                    "type": "string"
                },
                "expAccessToken": {
                    "type": "string"
                }
            }
        },
        "DTO.LoginResponse": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "type": "string"
                },
                "expAccessToken": {
                    "type": "string"
                },
                "expRefreshToken": {
                    "type": "string"
                },
                "refreshToken": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/DTO.UserResponse"
                }
            }
        },
        "DTO.UserResponse": {
            "type": "object",
            "required": [
                "firstName",
                "lastName"
            ],
            "properties": {
                "createdAt": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lastName": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "integer"
                }
            }
        },
        "swagger.AccessForbiddenResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/swagger.EmptyData"
                },
                "msg": {
                    "type": "string",
                    "example": "Sorry you don't have access to visit this page!"
                },
                "ok": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "swagger.EmptyData": {
            "type": "object"
        },
        "swagger.FailedLoginResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/swagger.EmptyData"
                },
                "msg": {
                    "type": "string",
                    "example": "No user found with entered credentials"
                },
                "ok": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "swagger.FailedResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/swagger.EmptyData"
                },
                "msg": {
                    "type": "string",
                    "example": "Error or warnnig message"
                },
                "ok": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "swagger.FailedValidationResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/swagger.validationErrors"
                },
                "msg": {
                    "type": "string",
                    "example": "Please review your entered data"
                },
                "ok": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "swagger.LoginResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/DTO.LoginResponse"
                },
                "msg": {
                    "type": "string",
                    "example": "Successful message"
                },
                "ok": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "swagger.NotFoundResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/swagger.EmptyData"
                },
                "msg": {
                    "type": "string",
                    "example": "404 not found!"
                },
                "ok": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "swagger.PaginateUsersData": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer",
                    "example": 10
                },
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/DTO.UserResponse"
                    }
                }
            }
        },
        "swagger.PingResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    },
                    "example": {
                        "pingpong": "🏓🏓🏓🏓🏓🏓"
                    }
                },
                "msg": {
                    "type": "string",
                    "example": "pong"
                },
                "ok": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "swagger.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/swagger.EmptyData"
                },
                "msg": {
                    "type": "string",
                    "example": "Successful message"
                },
                "ok": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "swagger.SuccessVerifyAccessTokenResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/DTO.AccessTokenRes"
                },
                "msg": {
                    "type": "string",
                    "example": "Successful message"
                },
                "ok": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "swagger.UnauthenticatedResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/swagger.EmptyData"
                },
                "msg": {
                    "type": "string",
                    "example": "You must login first!"
                },
                "ok": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "swagger.UsersListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/swagger.PaginateUsersData"
                },
                "msg": {
                    "type": "string",
                    "example": "Successful message"
                },
                "ok": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "swagger.validationErrors": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    },
                    "example": {
                        "field1": "This field is required",
                        "field2": "This field must be numeric"
                    }
                }
            }
        }
    }
}